Fix IndexOf in Utf8String using 64 bits instead of 32 bits

+++ FFXIVClientStructs/FFXIV/Client/System/String/Utf8String.cs
@@ -143,13 +143,13 @@ public unsafe partial struct Utf8String : ICreatable<Utf8String>, IDisposable, I
     public partial Utf8String* Replace(Utf8String* toFind, Utf8String* replacement);
 
     [MemberFunction("E8 ?? ?? ?? ?? B9 ?? ?? ?? ?? 45 84 ED")]
-    public partial int IndexOf(Utf8String* toFind, int startIdx = 0);
+    public partial ulong IndexOf(Utf8String* toFind, ulong startIdx = 0);
 
     [MemberFunction("48 89 5C 24 ?? 55 56 57 41 54 41 55 41 56 41 57 48 8D 6C 24 ?? 48 81 EC ?? ?? ?? ?? 48 8B 05 ?? ?? ?? ?? 48 33 C4 48 89 45 17 45 33 ED 41 BA")]
-    public partial int FindFirstOf(Utf8String* charsToFind, int startIdx, bool exclude = false);
+    public partial ulong FindFirstOf(Utf8String* charsToFind, ulong startIdx, bool exclude = false);
 
     [MemberFunction("48 89 5C 24 ?? 48 89 6C 24 ?? 48 89 74 24 ?? 57 41 54 41 55 41 56 41 57 48 83 EC 20 48 8B F9 45 0F B6 E9")]
-    public partial int FindLastOf(Utf8String* toFind, int startIdx, bool exclude = false);
+    public partial ulong FindLastOf(Utf8String* toFind, ulong startIdx, bool exclude = false);
 
     [MemberFunction("48 8B 01 0F B6 04")]
     public partial byte GetCharAt(ulong idx);
